divide into directories
