+++ cmd/go_todo/go.mod
@@ -1,0 +1,9 @@
+module go_todo
+
+go 1.14
+
+require (
+	github.com/gin-gonic/gin v1.6.3
+	github.com/jinzhu/gorm v1.9.14
+	github.com/joho/godotenv v1.3.0
+)
