--- docker-compose.yml
@@ -1,25 +1,0 @@
-#3系が最新らしい 書き方が異なる
-version: "3"
-services:
-  mysql: #ここの名前はなんでもよさそう
-    build:
-      context: .
-      dockerfile: ./mysql/Dockerfile
-    env_file:
-      - .env
-    volumes:
-      - ./mysql/db:/docker-entrypoint-initdb.d
-    environment:
-       MYSQL_ROOT_PASSWORD: ${DBPASS}
-  golang:
-    build:
-      context: .
-      dockerfile: ./go/Dockerfile
-    ports:
-      - "8080:8080"
-    depends_on:
-      - mysql #mysqlの方が先に起動される
-    stdin_open: true
-    tty: true # コンテナの起動永続化
-    volumes:
-      - ./:/go/src/go_todo
