--- mysql/Dockerfile
@@ -1,7 +1,0 @@
-FROM mysql
-#portを開ける
-EXPOSE 3306
-#設定ファイルをイメージ内にコピー
-#ADD ./my.cnf /etc/mysql/conf.d/my.cnf
-
-CMD ["mysqld"]
